+++ src/tsfile/encoding/bitpacking/LongPacker.go
@@ -26,12 +26,12 @@ import (
  */
 
 const (
-	NUM_OF_LONGS int = 8
+	NUM_OF_LONGS int32 = 8
 )
 
 type LongPacker struct {
 	//bit-width
-	BitWidth int
+	BitWidth int32
 }
 
 /**
@@ -41,17 +41,17 @@ type LongPacker struct {
  * @param offset - the offset of first Integer to be encoded
  * @param buf    - encoded bytes, buf size must be equal to ({@link LongPacker#NUM_OF_LONGS} * {@link LongPacker#width} / 8)
  */
-func (p *LongPacker) Pack8Values(values []int64, offset int, buf []byte) {
-	var bufIdx int = 0
-	var valueIdx int = offset
+func (p *LongPacker) Pack8Values(values []int64, offset int32, buf []byte) {
+	var bufIdx int32 = 0
+	var valueIdx int32 = offset
 	//remaining bits for the current unfinished Integer
-	var leftBit int = 0
+	var leftBit int32 = 0
 
 	for valueIdx < NUM_OF_LONGS+offset {
 		// buffer is used for saving 32 bits as a part of result
 		var buffer int64 = 0
 		// remaining size of bits in the 'buffer'
-		var leftSize int = 64
+		var leftSize int32 = 64
 
 		// encode the left bits of current Integer to 'buffer'
 		if leftBit > 0 {
@@ -93,13 +93,13 @@ func (p *LongPacker) Pack8Values(values []int64, offset int, buf []byte) {
  * @param offset - offset of first byte to be decoded in buf
  * @param values - decoded result , the length of 'values' should be @{link LongPacker#NUM_OF_LONGS}
  */
-func (p *LongPacker) Unpack8Values(buf []byte, offset int, values []int64) {
-	var byteIdx int = offset
-	var valueIdx int = 0
+func (p *LongPacker) Unpack8Values(buf []byte, offset int32, values []int64) {
+	var byteIdx int32 = offset
+	var valueIdx int32 = 0
 	//left bit(s) available for current byte in 'buf'
-	var leftBits int = 8
+	var leftBits int32 = 8
 	//bits that has been read for current long value which is to be decoded
-	var totalBits int = 0
+	var totalBits int32 = 0
 
 	//decode long value one by one
 	for valueIdx < 8 {
@@ -146,15 +146,15 @@ func (p *LongPacker) Unpack8Values(buf []byte, offset int, values []int64) {
  * @param length: length of bytes to be decoded in buf.
  * @param values: decoded result.
  */
-func (p *LongPacker) UnpackAllValues(buf []byte, length int, values []int64) {
-	var idx int = 0
-	var k int = 0
+func (p *LongPacker) UnpackAllValues(buf []byte, length int32, values []int64) {
+	var idx int32 = 0
+	var k int32 = 0
 	for idx < length {
 		tv := make([]int64, 8)
 		//decode 8 values one time, current result will be saved in the array named 'tv'
 		p.Unpack8Values(buf, idx, tv)
 
-		for i := 0; i < 8; i++ {
+		for i := int32(0); i < 8; i++ {
 			values[k+i] = tv[i]
 		}
 		idx += p.BitWidth
